Updated Dir Path in Server to accommodate Restructure
